+++ skrub/_to_float.py
@@ -8,8 +8,8 @@ class ToFloat(SingleColumnTransformer):
     """
     Convert a column to 32-bit floating-point numbers.
 
-    No conversion is attempted if the column has a datetime or categorical
-    dtype; a ``RejectColumn`` exception is raised.
+    No conversion is attempted if the column has a datetime, duration or
+    categorical dtype; a ``RejectColumn`` exception is raised.
 
     Otherwise, we attempt to convert the column to float32. If the conversion
     fails the column is rejected (a ``RejectColumn`` exception is raised).
@@ -185,7 +185,11 @@ class ToFloat(SingleColumnTransformer):
         """
         del y
         self.all_outputs_ = [sbd.name(column)]
-        if sbd.is_any_date(column) or sbd.is_categorical(column):
+        if (
+            sbd.is_any_date(column)
+            or sbd.is_duration(column)
+            or sbd.is_categorical(column)
+        ):
             raise RejectColumn(
                 f"Refusing to cast column {sbd.name(column)!r} "
                 f"with dtype '{sbd.dtype(column)}' to numbers."